[fix]: Resolve ci errors and optimize code
