+++ packages/sqle/src/page/ProjectOverview/component/DataSourcePerformance/index.tsx
@@ -3,59 +3,20 @@ import { useTranslation } from 'react-i18next';
 import { useState } from 'react';
 import useChatsDataByAPI from '../../hooks/useChatsDataByAPI';
 import { useCurrentProject } from '@actiontech/shared/lib/global';
-import { Bar, BarConfig, Tooltip } from '@ant-design/plots';
+import { Bar, BarConfig } from '@ant-design/plots';
 import { useChangeTheme } from '@actiontech/shared/lib/hooks';
 import useThemeStyleData from '../../../../hooks/useThemeStyleData';
 import ChartWrapper from '../../../../components/ChartCom/ChartWrapper';
-import ChartTooltip from '../../../../components/ChartCom/ChartTooltip';
-import { SharedTheme } from '@actiontech/shared/lib/types/theme.type';
-import { SqleTheme } from '../../../../types/theme.type';
 import sqlOptimization from '@actiontech/shared/lib/api/sqle/service/sql_optimization';
 import { IDBPerformanceImproveOverview } from '@actiontech/shared/lib/api/sqle/service/common';
-import { t } from '../../../../locale';
-import { floatToPercent } from '@actiontech/shared/lib/utils/Math';
+import {
+  defaultItemKey,
+  labelFormatter,
+  renderTooltipCustomContent,
+  renderTooltipFormatter
+} from './index.data';
 
 const limitDataLength = 8;
-
-const defaultItemKey = 'performance-default-key';
-
-const renderTooltipFormatter: Tooltip['formatter'] = (item) => {
-  return {
-    name: item.instance_name,
-    value: item.avg_performance_improve
-  };
-};
-
-const renderTooltipCustomContent = (
-  dataSource: Array<{ name: string; value: number }>,
-  sqleTheme: SqleTheme,
-  sharedTheme: SharedTheme
-) => {
-  const data = dataSource[0];
-  if (!data) return null;
-  if (data.name.startsWith(defaultItemKey)) return null;
-  return (
-    <ChartTooltip
-      titleData={{
-        dotColor:
-          sqleTheme.projectOverview.DataSourcePerformance.toolTip.dotColor,
-        text: data.name
-      }}
-      listData={[
-        {
-          label: t(
-            'projectManage.projectOverview.dataSourcePerformance.toolTip.label'
-          ),
-          value: `${floatToPercent(data.value)}%`
-        }
-      ]}
-      sharedTheme={sharedTheme}
-    />
-  );
-};
-
-const labelFormatter = (value: string) =>
-  value.startsWith(defaultItemKey) ? '' : value;
 
 const DataSourcePerformance = () => {
   const { t } = useTranslation();
@@ -74,12 +35,12 @@ const DataSourcePerformance = () => {
     },
     {
       onSuccess: (res) => {
-        const { data = [] } = res.data;
+        const list = res.data.data ?? [];
         const emptyData: IDBPerformanceImproveOverview[] = [];
-        if (data.length) {
+        if (list.length) {
           const emptyDataLength =
-            limitDataLength - data.length > 0
-              ? limitDataLength - data.length
+            limitDataLength - list.length > 0
+              ? limitDataLength - list.length
               : 0;
           if (emptyDataLength) {
             for (let i = 0; i < emptyDataLength; i++) {
@@ -90,7 +51,7 @@ const DataSourcePerformance = () => {
             }
           }
         }
-        setData(data.concat(emptyData));
+        setData(list.concat(emptyData));
       }
     }
   );
@@ -165,10 +126,3 @@ const DataSourcePerformance = () => {
 };
 
 export default DataSourcePerformance;
-
-export {
-  renderTooltipFormatter,
-  renderTooltipCustomContent,
-  defaultItemKey,
-  labelFormatter
-};
